Adds Boss Animator FSM and Manager Script
Adds the FSM for the Mad Mantis via an Animator Controller, considering its first stage with an idle and two attacks, a second stage with the attacks and an extra jump state, and a third stage with both attacks and the flying state.
Adds a Manager script handling the boss' health, stage change and attacks and jump cooldown.
Adds a script for each idle stage to begin attacking when reaching the new boss fight stage.
TODO: implement jump, fly and transitions' visual effects