Adds enraged logic and shader for visuals
Adds a script to "enrage" the mad mantis in the animator and a shader manager to control the values and states, as well as a shader graph to color it gradually red and make it wavy during the enraging animation.
Adds the URP package to enable using shadergraph with sprites.