Adds new visuals to mantis and enhance its attacks
Adds the new sprites and animations to the mad mantis prefab and animator.
Changes the bullet spawn of its attacks to be made at specific animation's frames, for better visuals.
Adds normal maps to all sprites, and make its shader be a "lit" shader.
Changes the "boss" scene to have the new tileset and some illumination, better representing the final setup of the scene.