Updates Mad Mantis animations and controller
Updates Mad Mantis animations & animator to new animations and also to be more fluid. Updates bullets' spawn points and spawn timing to be more precise to the animations' movements.
Updates mad mantis controller to separate the animations' start from the bullets' spawn. Now, the bullets are spawned via animation event.
Adds animation for each attack to spawn a bullet in a higher or lower position, and a logic to give 50% chance for each position.