Adds effects and controller for transition into last mantis stage
Adds post processing effects and a shader. They are triggered when entering the last stage of the mantis boss fight. The transition now happens in two stages: first, all coroutines are stopped and the wave + chromatic aberration + negative color effects are called. Then, a vignete and hue shift are added until the end of the fight.
Changes some renderer settings to support the wave effect.