Changes camera and spawn to handle checkpoints
Changes spawn manager to hold the value of a new collider, from the current checkpoint region. Initializing player's health uses this value to move the player's and camera's position to the new point and use the collider as the new camera's bounds.
Adds some minor refactorings.
Changes the camerarelated prefab to not hold the camera regions. They are now part of a new object (not destroyable on load) that holds both spawn points and camera regions, so they are maintained over game-overs.